+++ src/errors/base-elysia-error.ts
@@ -1,7 +1,0 @@
-import { ErrorCode } from '../types'
-
-export class ElysiaError extends Error {
-	constructor(message: ErrorCode, opts?: ErrorOptions) {
-		super(message, opts)
-	}
-}