+++ src/errors/index.ts
@@ -1,2 +1,0 @@
-export * from './base-elysia-error'
-export * from './validation-error'